Added BundleManagementManagedBean (WIP), bundleManagement.xhtml (WIP), changed erroneous delete method in BundleSessionBean
Added BundleManagementManagedBean.java which is currently not working as intended
java.sql.SQLSyntaxErrorException: Unknown column 't1.ISONDISPLAY' in 'field list' is shown when trying to access page bundleManagement.xhtml
bundleManagement.xhtml added which is currently not working as intended
bundleManagement.xhtml is not currently exposed in the main menu. It can be accessed by direct url path.